copies cpu.py, and resets it so I can start over
