+++ OLD_ls8/examples/call.ls8
@@ -1,0 +1,31 @@
+10000010 # LDI R1,MULT2PRINT
+00000001
+00011000
+10000010 # LDI R0,10
+00000000
+00001010
+01010000 # CALL R1
+00000001
+10000010 # LDI R0,15
+00000000
+00001111
+01010000 # CALL R1
+00000001
+10000010 # LDI R0,18
+00000000
+00010010
+01010000 # CALL R1
+00000001
+10000010 # LDI R0,30
+00000000
+00011110
+01010000 # CALL R1
+00000001
+00000001 # HLT
+# MULT2PRINT (address 24):
+10100000 # ADD R0,R0
+00000000
+00000000
+01000111 # PRN R0
+00000000
+00010001 # RET
