+++ OLD_ls8/examples/interrupts.ls8
@@ -1,0 +1,25 @@
+10000010 # LDI R0,0XF8
+00000000
+11111000
+10000010 # LDI R1,INTHANDLER
+00000001
+00010001
+10000100 # ST R0,R1
+00000000
+00000001
+10000010 # LDI R5,1
+00000101
+00000001
+10000010 # LDI R0,LOOP
+00000000
+00001111
+# LOOP (address 15):
+01010100 # JMP R0
+00000000
+# INTHANDLER (address 17):
+10000010 # LDI R0,65
+00000000
+01000001
+01001000 # PRA R0
+00000000
+00010011 # IRET
