+++ OLD_ls8/examples/keyboard.ls8
@@ -1,0 +1,28 @@
+10000010 # LDI R0,0XF9
+00000000
+11111001
+10000010 # LDI R1,INTHANDLER
+00000001
+00010001
+10000100 # ST R0,R1
+00000000
+00000001
+10000010 # LDI R5,2
+00000101
+00000010
+10000010 # LDI R0,LOOP
+00000000
+00001111
+# LOOP (address 15):
+01010100 # JMP R0
+00000000
+# INTHANDLER (address 17):
+10000010 # LDI R0,0XF4
+00000000
+11110100
+10000011 # LD R1,R0
+00000001
+00000000
+01001000 # PRA R1
+00000001
+00010011 # IRET
