+++ OLD_ls8/examples/mult.ls8
@@ -1,0 +1,12 @@
+10000010 # LDI R0,8
+00000000
+00001000
+10000010 # LDI R1,9
+00000001
+00001001
+10100010 # MUL R0,R1
+00000000
+00000001
+01000111 # PRN R0
+00000000
+00000001 # HLT
