+++ OLD_ls8/examples/printstr.ls8
@@ -1,0 +1,56 @@
+10000010 # LDI R0,HELLO
+00000000
+00100110
+10000010 # LDI R1,14
+00000001
+00001110
+10000010 # LDI R2,PRINTSTR
+00000010
+00001100
+01010000 # CALL R2
+00000010
+00000001 # HLT
+# PRINTSTR (address 12):
+10000010 # LDI R2,0
+00000010
+00000000
+# PRINTSTRLOOP (address 15):
+10100111 # CMP R1,R2
+00000001
+00000010
+10000010 # LDI R3,PRINTSTREND
+00000011
+00100101
+01010101 # JEQ R3
+00000011
+10000011 # LD R3,R0
+00000011
+00000000
+01001000 # PRA R3
+00000011
+01100101 # INC R0
+00000000
+01100110 # DEC R1
+00000001
+10000010 # LDI R3,PRINTSTRLOOP
+00000011
+00001111
+01010100 # JMP R3
+00000011
+# PRINTSTREND (address 37):
+00010001 # RET
+# HELLO (address 38):
+01001000 # H
+01100101 # e
+01101100 # l
+01101100 # l
+01101111 # o
+00101100 # ,
+00100000 # [space]
+01110111 # w
+01101111 # o
+01110010 # r
+01101100 # l
+01100100 # d
+00100001 # !
+00001010 # 0x0a
