+++ OLD_ls8/examples/stack.ls8
@@ -1,0 +1,31 @@
+10000010 # LDI R0,1
+00000000
+00000001
+10000010 # LDI R1,2
+00000001
+00000010
+01000101 # PUSH R0
+00000000
+01000101 # PUSH R1
+00000001
+10000010 # LDI R0,3
+00000000
+00000011
+01000110 # POP R0
+00000000
+01000111 # PRN R0
+00000000
+10000010 # LDI R0,4
+00000000
+00000100
+01000101 # PUSH R0
+00000000
+01000110 # POP R2
+00000010
+01000110 # POP R1
+00000001
+01000111 # PRN R2
+00000010
+01000111 # PRN R1
+00000001
+00000001 # HLT
