+++ OLD_ls8/examples/stackoverflow.ls8
@@ -1,0 +1,19 @@
+10000010 # LDI R0,0
+00000000
+00000000
+10000010 # LDI R1,1
+00000001
+00000001
+10000010 # LDI R3,LOOP
+00000011
+00001001
+# LOOP (address 9):
+01000111 # PRN R0
+00000000
+10100000 # ADD R0,R1
+00000000
+00000001
+01000101 # PUSH R0
+00000000
+01010100 # JMP R3
+00000011
