+++ OLD_ls8/ls8.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+
+"""Main."""
+
+import sys
+from cpu import *
+
+cpu = CPU()
+
+cpu.load()
+cpu.run()
